Change overlap-sum to use accumulator.
The OverlapSum block computes a moving average. Rather than perform the
entire sum for each output, use a shift register to subtract the sample
that is leaving the window and only add the sample that is entering the
window. Changing this exposed some other weirdness in how the
memory-based shift register worked, and some tests were also updated.